Their Groups Code
I thought we had alot more time to do this, but I was wrong. Their code works almost the same, but im gonna make some adjustments. It's setup for how I want to continue with the project so we'll use this as a base it'll be faster this way. So I'm just gonna give you what they had in the previous assignment to trace and work with. I'm Gonna add the Changes I've made to it after but I want you guys to have something to work with.  Upload your changes to Branches and we'll compile the finshed product from there